refactor(core): decompose worktree god-file into submodules
worktree.rs was 1982 lines (god_file) with decomposition seams already
present as inline mod blocks (types, store_ops, queue_ops) plus a large
test module and the thin top-level pub fn API.

Move each inline module into its own file under worktree/, keeping the
imports, the types re-export, the top-level public API fns (create,
adopt, list, status, remove, cleanup, queue_create, ...), and the
`use store_ops::*;` / `use queue_ops::*;` bring-into-scope statements in
mod.rs. Public API is byte-identical and no logic changed.

Largest resulting file is 725 lines (store_ops); all well under the
1500-line god_file threshold.

Clears the worktree.rs god_file finding.
